Django/requirements.txt
asgiref==3.5.2
backcall==0.2.0
backports.zoneinfo==0.2.1
charset-normalizer==2.1.1
colorama==0.4.5
decorator==5.1.1
Django==3.2.15
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.4
ipdb==0.13.9
ipython==7.34.0
jedi==0.18.1
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.31
Pygments==2.13.0
PyJWT==2.4.0
pytz==2022.2.1
razorpay==1.3.0
requests==2.28.1
sqlparse==0.4.2
toml==0.10.2
traitlets==5.3.0
typing_extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
wcwidth==0.2.5
whitenoise==6.2.0
wincertstore==0.2
